p4i integration Integrate 235109 from 4.2 to main:
SH_ScrollBar_StopMouseOverSlider was obsoleted by
SH_Slider_StopMouseOverSlider. Reorder the enums so that the primary
one doesn't depend on the old one. We need this from Jambi so that we
can kill the obsoleted one. Since they are sequentially defined
this won't change the values so its safe... (tm)

[git-p4: depot-paths = "//depot/qt/main/": change = 235111]