p4i integration Integrate 235539 from 4.2 to main: Fix crash in Mac accessibility when handling QObject destruction.
QInterfaceItem::valid() should also check if it's QAccessibleInterface is still valid, not
just that the pointer is not null. 

In addition, the QAccessibleInterface is already invalid  when get the objectDestroyed 
call for it's object. The reason for this is that QAccessibleObject stores the object
pointer using a QPointer, which also listens on the destroyed signal and resets
the pointer before we can get to it. 

This is unfortunate since we need the object pointer in the qHash function to look up 
the Mac-side AXUIElements. The solution is to store a copy of the qobject pointer in 
each QInterfaceItem.

[git-p4: depot-paths = "//depot/qt/main/": change = 235540]