p4i integration Integrate 236336 from 4.2 to main: Fix QSizeGrip resizing bug on Mac: Resizing a window while a popup menu was open would cause it to jump to a very large size.
When you left-click to close a popup menu on in native Mac applications, that 
click is a "dead" mouse click and does not cause any interaction with the application.

The way this is currently implemented in Qt is that we don't get the pressed event, but
do get the following move and release events. This can break widgets that assume that
you always get mouse pressed before move or release, and QSizeGrip is one of those 
widgets.

Fix this by checking if we got the pressed event before trying to resize.

RevBy:    Trenton

[git-p4: depot-paths = "//depot/qt/main/": change = 236337]