Fixes:    flaws in asm implementations of atomics on several architectures Task:     126321 Details:
apply patch from dirk mueller (note that i am not including the
gcc/icc sync builtins in this change):

- the architecture specific version as a q_atomic_set_ptr()
function that casts a void* pointer to void** and dereferences it.
this introduces a strict aliasing bug with optimizing compilers. 

- the asm implementations for several architectures contain flaws: 
* memory barriers are partially forgotten sometimes
* they implement the ptr variant via the integer variant, which
introduces yet another strict aliasing bug (and this bug has been
observed in practice with gcc 4.2). 
* register clobbers are incorrect, causing optimizing compilers
to produce spurious failure. 

[git-p4: depot-paths = "//depot/qt/main/": change = 237963]