p4i integration Integrate 242074 from 4.2 to main: Fixes:    A thread posting lots of custom events on Unix may cause the event loop of the receiver to not wake up Task:     138597 RevBy:    morten, janerik, customer AutoTest: not testable :/ Details:
the unix event dispatcher uses an atomic variable as a semaphore
to choke the number of bytes written to the wakeup pipe. apparently,
we get into the situation where the receiver thread has emptied the
pipe, assigned zero to d->wakeUps (but it hasn't happened in the MMU
yet), and has gone into select().

when the sender thread calls wakeUp(), d->wakeUps.testAndSet() fails,
meaning we don't write to the pipe, and a deadlock results. :/

[git-p4: depot-paths = "//depot/qt/main/": change = 242075]