Make the the native mouse event tests pass, and also make clicking on the title bar not generate a (client area) mouse release event.
Details:
Keep track of whether a  mouse press was in the client area or not when
setting the qt_button_down widget, use this information when deciding to
generate a non-client area mouse event or a normal mouse event.

Don't listen for mouse up events in the window event handler, leave this
for the application event handler. We were getting multiple non-client
area mouse release events when clicking on the title bar, I believe this
was because we return "not handled" after handling the event so that
the standard/default event handler will get it. Since we had installed 
two event handlers for it we were emitting mouse events twice.

[git-p4: depot-paths = "//depot/qt/main/": change = 245264]