p4i integration Integrate 255128 from 4.3 to main: Fixes:    Low priority threads hogging CPU Task:     none RevBy:    sroedal AutoTest: none Details:
On Linux, using SCHED_RR preempts all other processes and threads, regardless of how
the scheduling paramter is set, see http://www.die.net/doc/linux/man/man2/sched_setscheduler.2.html

To me, this seems to be in conflict with POSIX, but POSIX is so abmiguous that this
can be debated. Regardless, we can't have "low" priority threads hogging the CPU and
starving the rest of the system.

[git-p4: depot-paths = "//depot/qt/main/": change = 255129]