p4i integration Integrate 259462 from 4.3 to main: Compile on Mac. Change 258635 makes checkReceiverThread() an inlined empty function in release mode. This causes link errors when using frameworks, since QtGui built in debug mode will link against QtCore in release mode (this is how frameworks work, the release version is selected by default if the framework contains both a debug and release library).
So what happens is that the debug QtGui expects to find checkReceiverThread()
in the QtCore library but fails since it's inline in the release version of QtCore.
The fix is to make sure the function is always compiled into QtCore when building
with frameworks on Mac, and then only call it in debug mode.

RevBy:Brad

[git-p4: depot-paths = "//depot/qt/main/": change = 259463]