p4i integration Integrate 259685 from 4.3 to main: AutoTest: included Details:  From Thiago:
This patch contains a set of bugfixes for QDBusArgument, as well as
adding one feature: allowing one to create and use QDBusArgument
objects. Currently, Qt only allows QDBusArguments to be used in the
operator<< and operator>> functions. With this patch, you can marshall
your own stuff.

Richard Dale requires this for proper C# and Ruby bindings.

The bug being fixed is seen in the last file in the patch: the object
didn't detach properly, so the DBusMessageIter iterator was
advanced. If you tried to read the same QDBusArgument twice (for
instance, if you had two slots connected to the same signal), the
second would fail with weird libdbus errors (read past the end of
message).

This implements proper implicit detaching.

[git-p4: depot-paths = "//depot/qt/main/": change = 259758]