p4i integration Integrate 271058 from 4.3 to main: Fixes: undo after Backspace (regression from 4.2.x)        Ctrl-x when there is no selection (regression from even earlier releases) Task: blackteam RevBy: jesper AutoTest: included (for the Undo stuff) Details:
Change 252536 broke Undo behavior when Backspace was used. Typing in
ABC, Backspace, Ctrl-z with give an empty lineedit, not ABC as
expected. I changed it to only insert a SetSelection command when there
actually WAS a selection to remember, if not no SetSelection will be
entered.

Also noticed that Cut (Ctrl-X) worked, regardless if anything was
selected or not. This is a regression, made it call cut() as it should
anyways.

[git-p4: depot-paths = "//depot/qt/main/": change = 271059]