p4i integration Integrate 271150 from 4.3 to main: Fixes:    scrolling + CachedBackground in QGraphicsView Task:     170619 RevBy:    Eskil
Details: The code for handling scrolling of cached backgrounds did not
work on neither windows or mac as CompositionMode_Source is not
supported on Mac (at all) and not for this pixmap format on
Windows. Replaced this code with a fill(Qt::transparent) instead.

The fillrect code in scrollContentsBy() was wrong as it fills a a
rectangle with the wrong brush origin and all of this area will be
painted over again by either the expose event or the drawPixmap
scrolling on the following line.

The pixel offsetting put into backgroundPixmapExposed is also
wrong. There is no reason why thy y and the x should be treated
differently.

Finally... The paintEvent for the background draws the background by
iterating through all the rects in the region. It is then evident that
none of these rects overlap, otherwise alpha pixels will accumulate. We
set the actual cliprect prior to setting any transformation, thus
enforcing that we only repaint the specified areas.

[git-p4: depot-paths = "//depot/qt/main/": change = 271151]