Fixes:    System font changes ignored on windows Task:     172757 RevBy:    prasanth Details: This fixes a regression following change 135417. Though it is a regression, we decided it wasnt important enough to go into 4.3.2.
Previously we would trigger a font change following a color change.
This works because they are allways sent at the same time when a user
changes options in the appearence panel on Windows. However after
this change we did not recieve font changes while the application was running
any more. Note that WA_FONTCHANGE doesnt mean system font change, but
that a font has been added to the system. A WM_SETTINGCHANGE with
SPI_SETICONTITLELOGFONT seems to be the most appropriate message for this.

[git-p4: depot-paths = "//depot/qt/main/": change = 274298]