p4i integration Integrate 275132 from 4.3 to main: Fixes:    performance regression in QLayout. Task:     none RevBy:    Thiago Details:
We can't use QMetaObject::invokeMethod in places that are 
performance critical. It is a lot better to just get the 
index of the method to call once and then use qt_metacall directly.

This cuts down the time spend in the layout by around 40%.
Basically QStyle::layoutSpacing() is now a factor 50
faster :)

[git-p4: depot-paths = "//depot/qt/main/": change = 275133]