p4i integration Integrate 275565 from 4.3 to main: Fixes:    Fix crash in the the OpenGL window-change update mechanism. RevBy:    Trond ManualTest: 275558 Details: (This change is related to 275216 and 271933.)
The crash was caused by a stale widget pointer in the 
kEventQtRequestWindowChange handler, we were getting window change events
long after the widget had been deleted.

So, instead of posting a kEventQtRequestWindowChange  event to the carbon
event queue and then sending a MacGLWindowChange event to the right widget
in its handler, we post a MacGLWindowChange event directly to the widget that
needs the update.

[git-p4: depot-paths = "//depot/qt/main/": change = 275566]